feat: add helper copyFileToDirectory()
